pyproject.toml
[tool.poetry]
name = "bicycle-rental"
version = "0.1.0"
description = ""
authors = ["Timurxboy <timurxboy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
setuptools = "^70.1.1"
drf-spectacular = {extras = ["sidecar"], version = "^0.27.1"}
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
drf-spectacular-sidecar = "2024.5.1"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
